New script to generate pdf documentation.
This is a convenience script to generate the documentation in pdf.
The command line to generate pdf is long, so I added this script
as a convenience. It might be benificial to others.
Instead of having to generate this manually, this script allows
the entire documentation to be generated with one command, or
simply an entire subdirectory or just a single file.