Update mkdocs to 1.3.0 and use new material theme
Provide an update to the latest stable version of mkdocs. This change
require to many modification of an outdates theme, for that reason a new
theme "Material" is included.

The new theme provide extensions to handle notes and other admonitions.
For that reason the class="well" html code is removed and replaced by
the new prefix (extension method).

References: #226