Use mkdocs material theme
* Update mkdocs to 1.3.0 and use new material theme

Provide an update to the latest stable version of mkdocs. This change
require to many modification of an outdates theme, for that reason a new
theme "Material" is included.

The new theme provide extensions to handle notes and other admonitions.
For that reason the class="well" html code is removed and replaced by
the new prefix (extension method).

References: #226

* Remove OpenIndiana Logo from all sub-pages

The logo should no longer be part of the sub-pages or pages because it's
part of the header. This is required to make the documentation more
readable for the user.

* Update index page overview

Add css code to create grid overview based on the html and markdown
code. Update copyright to the current year (2022).

* Keep div elements which are used by Pandoc to detect note blocks

* Adjust PDF script and config to be compatible with Markdown format changes

* Fix bug with -d option of PDF script
Also remove redundent check

* Remove lines with tailing spaces
These do not have any effect on formatting and are flagged as an error by Markdown Lint

* Fix linting errors
Mostly minor trailing spaces or blank line. Only major change is a table which needed updating to a compatible format.

* Update MkDocs Material version and use new features

* Fix broken links

* Remove deprecated config value
'The configuration option google_analytics has been deprecated and will be removed in a future release of MkDocs'

* Fix HCL links

* Add config for nested code block rendering

* Remove duplication from page titles

* Update docs contrib for Material theme

* Use frozen pip versions, Use verbose makepdf and strict mkdocs in CI

---------

Co-authored-by: JMadgwick <32958136+JMadgwick@users.noreply.github.com>